lesson14

+++ os_work.py
@@ -1,8 +1,8 @@
 import os
 import shutil
 
-def os_work():
 
+def os_work():
     while True:
 
         print('1. создать папку')
@@ -15,61 +15,56 @@ def os_work():
         print('8. смена директории')
         print('9. вернуться в главное меню')
 
-        choice =  input('Выберите пункт меню ')
+        choice = input('Выберите пункт меню ')
 
         work_path = os.getcwd()
         # print('рабочий путь',work_path)
 
         if choice == '1':
-            name_new = work_path+'/'+input('какую папку будем создавать? ')
-            if not os.path.exists(name_new):
-                os.mkdir(name_new)
-            else:
-                print('такая папка уже существует')
+            name_new = work_path + '/' + input('какую папку будем создавать? ')
+            os.mkdir(name_new) if not os.path.exists(name_new) else print('такая папка уже существует')
 
         if choice == '2':
             print(os.listdir(work_path))
             name_new = work_path + '/' + input('что будем удалять? ')
-            if os.path.exists(name_new):
-                os.rmdir(name_new)
-            else:
-                print('такой папки не существует')
-
+            os.rmdir(name_new) if os.path.exists(name_new) else print('такой папки не существует')
 
         if choice == '3':
             print(os.listdir(work_path))
-            name_old =  input('что копируем? имя + расширение')
+            name_old = input('что копируем? имя + расширение')
 
-            if os.path.exists(name_old):
-                name_new = input('как назовем? имя + расширение')
-                shutil.copyfile(work_path+'/'+name_old, work_path+'/'+name_new)
-            else:
+            name_new = input('как назовем? имя + расширение')
+            path_and_file_old = work_path + '/' + name_old
+            path_and_file_new = work_path + '/' + name_new
+
+            try:
+                shutil.copyfile(path_and_file_old, path_and_file_new)
+
+            except FileNotFoundError:
                 print('такого файла/папки не существует')
 
         if choice == '4':
-            print(os.listdir(work_path))   # посмотреть содержимое директории
+            print(os.listdir(work_path))  # посмотреть содержимое директории
 
         if choice == '5':
             for dirname in os.listdir(work_path):
-                if os.path.isdir(work_path+'/'+dirname):
-                    print("Каталог:", os.path.join(work_path+'/'+dirname))
-
+                temp_path = work_path + '/' + dirname
+                if os.path.isdir(temp_path):
+                    print("Каталог:", os.path.join(temp_path))
 
         if choice == '6':
-            # посмотреть фйлы
+            # посмотреть файлы
             for filename in os.listdir(work_path):
-                if os.path.isfile(work_path+'/'+filename):
-                    print("Каталог:", os.path.join(work_path+'/'+filename))
+                if os.path.isfile(work_path + '/' + filename):
+                    print("Каталог:", os.path.join(work_path + '/' + filename))
 
         if choice == '7':
             print('окружение', os.environ)
 
         if choice == '8':
             print('текущая директория', work_path)
-            os.chdir(input('введите новый путь' ))
-            print(os.getcwd() )
+            os.chdir(input('введите новый путь'))
+            print(os.getcwd())
 
         if choice == '9':
             break
-
-
